COC-5 Refactor: @/componenets structure
