+++ apps/native/components/custom/list-item-card-thumbnail/index.tsx
