+++ apps/native/components/custom/navigation/TabBarIcon.tsx
