+++ apps/native/components/custom/text/discount-period/index.tsx
