+++ apps/native/components/custom/text/discount-rate/index.tsx
